requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
networkx>=3.0,<4.0
scikit-learn>=1.3.0,<2.0.0
tqdm>=4.65.0,<5.0.0
wikipedia>=1.4.0,<2.0.0

# Machine Learning
sentence-transformers>=2.2.0,<3.0.0
bertopic>=0.15.0,<1.0.0
umap-learn>=0.5.3,<1.0.0
hdbscan>=0.8.29,<1.0.0

# Vector Search
faiss-cpu>=1.7.2  # Use faiss-gpu if you have CUDA

# Caching and Optimization
redis>=4.5.1,<5.0.0
msgpack>=1.0.5,<2.0.0
xxhash>=3.2.0,<4.0.0

# Development
pytest>=7.0.0,<8.0.0
black>=22.0.0,<23.0.0
flake8>=6.0.0,<7.0.0
jupyter>=1.0.0,<2.0.0
